% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_data.R
\docType{data}
\name{data_prev_rec_alt}
\alias{data_prev_rec_alt}
\title{Intervention: prevention of recurrent depression (alternative)}
\format{
Same structure as \code{data_prev_rec_alt}.
}
\usage{
data(data_prev_rec_alt)
}
\description{
Alternative scenario intervention parameters for the prevention of recurrent
depression. Structure matches the base dataset. Values can be adjusted to reflect alternative modelling
assumptions. In this dataset, the same numbers are provided as in the base case.
}
\keyword{datasets}
