% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_data.R
\docType{data}
\name{data_prev_sub_alt}
\alias{data_prev_sub_alt}
\title{Intervention: prevention of sub-clinical depression (alternative)}
\format{
A data frame with the same columns as
\code{data_prev_sub_alt}.
}
\usage{
data(data_prev_sub_alt)
}
\description{
This dataset contains alternative scenario intervention parameters for the
prevention of sub-clinical depression in the DepMod model. The structure is
identical to the base dataset but can represent alternative modelling
assumptions. In this dataset, the same numbers are provided as in the base case.
}
\keyword{datasets}
