% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_data.R
\docType{data}
\name{data_tr_mild_alt}
\alias{data_tr_mild_alt}
\title{Intervention: treatment of mild depression (alternative)}
\format{
Same structure as \code{data_tr_mild_base}.
}
\usage{
data(data_tr_mild_alt)
}
\description{
Alternative scenario parameters for the treatment of mild depression. The
structure matches the base dataset but values can be adjusted to reflect alternative modelling
assumptions. In this dataset, the same numbers are provided as in the base case.
}
\keyword{datasets}
