% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_data.R
\docType{data}
\name{data_tr_mod_alt}
\alias{data_tr_mod_alt}
\title{Intervention: treatment of moderate depression (alternative)}
\format{
Same structure as \code{data_tr_mod_base}.
}
\usage{
data(data_tr_mod_alt)
}
\description{
Alternative scenario parameters for the treatment of moderate depression,
structurally identical to the base dataset. Values can be adjusted to reflect alternative modelling
assumptions. In this dataset, the same numbers are provided as in the base case.
}
\keyword{datasets}
