% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_data.R
\docType{data}
\name{data_tr_sev_alt}
\alias{data_tr_sev_alt}
\title{Intervention: treatment of severe depression (alternative)}
\format{
Same structure as \code{data_tr_sev_alt}.
}
\usage{
data(data_tr_sev_alt)
}
\description{
Alternative intervention parameters for the treatment of severe depression
episodes. Structure matches the base dataset. Values can be adjusted to reflect alternative modelling
assumptions. In this dataset, the same numbers are provided as in the base case.
}
\keyword{datasets}
