% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_data.R
\docType{data}
\name{data_tr_sev_base}
\alias{data_tr_sev_base}
\title{Intervention: treatment of severe depression (base)}
\format{
Same structure as \code{data_tr_sev_base}.
}
\usage{
data(data_tr_sev_base)
}
\description{
Baseline intervention parameters for the treatment of severe depression
episodes. Includes coverage, adherence, effectiveness, sample size, and
healthcare costs.
}
\keyword{datasets}
